:recycle: コンテキストと状態をコンポーネントから分離
