+++ src/components/test_use_context_2/ChildSecond.tsx
@@ -1,6 +1,6 @@
 import { FC, useContext } from 'react';
 
-import { ParentContext } from './Parent';
+import { ParentContext } from './ContextProvider';
 
 const ChildSecond: FC = () => {
   // 3. Context Valueの取得
