+++ src/components/test_use_context_2/Parent.tsx
@@ -1,22 +1,16 @@
-import { Dispatch, FC, SetStateAction, createContext, useState } from 'react';
+import { FC } from 'react';
 
 import Child from './Child';
 import ChildSecond from './ChildSecond';
-
-// 1. contextの作成
-type ParentContextType = [number, Dispatch<SetStateAction<number>>];
-export const ParentContext = createContext<ParentContextType>({} as never);
+import ContextProvider from './ContextProvider';
 
 const Parent: FC = () => {
-  const [state, setState] = useState(0);
-
   return (
-    // 2. Provider valueに、状態を更新関数を指定
-    <ParentContext.Provider value={[state, setState]}>
+    <ContextProvider>
       <h1>Parent</h1>
       <Child />
       <ChildSecond />
-    </ParentContext.Provider>
+    </ContextProvider>
   );
 };
